requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
cryptography==42.0.2
paramiko==3.4.0
sqlalchemy==2.0.25
pydantic==2.5.3